webgpu_traits::messages::recv

Enum WebGPURequest

pub enum WebGPURequest {

    SetImageKey {
        context_id: WebGPUContextId,
        image_key: ImageKey,
    },
    BufferMapAsync {
        callback: GenericCallback<Result<Mapping, BufferAccessError>>,
        buffer_id: BufferId,
        device_id: DeviceId,
        host_map: HostMap,
        offset: u64,
        size: Option<u64>,
    },
    CommandEncoderFinish {
        command_encoder_id: CommandEncoderId,
        device_id: DeviceId,
        desc: CommandBufferDescriptor<Label<'static>>,
        command_buffer_id: CommandBufferId,
    },
    CopyBufferToBuffer {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        source_id: BufferId,
        source_offset: BufferAddress,
        destination_id: BufferId,
        destination_offset: BufferAddress,
        size: BufferAddress,
    },
    CopyBufferToTexture {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        source: TexelCopyBufferInfo,
        destination: TexelCopyTextureInfo,
        copy_size: Extent3d,
    },
    CopyTextureToBuffer {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        source: TexelCopyTextureInfo,
        destination: TexelCopyBufferInfo,
        copy_size: Extent3d,
    },
    CopyTextureToTexture {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        source: TexelCopyTextureInfo,
        destination: TexelCopyTextureInfo,
        copy_size: Extent3d,
    },
    CopyExternalImageToTexture {
        device_id: DeviceId,
        queue_id: QueueId,
        usable_source: Option<SharedSnapshot>,
        destination: TexelCopyTextureInfo,
        dest_tex_descriptor: TextureDescriptor<'static>,
        copy_size: Extent3d,
    },
    CommandEncoderPushDebugGroup {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        label: String,
    },
    CommandEncoderPopDebugGroup {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
    },
    CommandEncoderInsertDebugMarker {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        label: String,
    },
    CreateBindGroup {
        device_id: DeviceId,
        bind_group_id: BindGroupId,
        descriptor: BindGroupDescriptor<'static>,
    },
    CreateBindGroupLayout {
        device_id: DeviceId,
        bind_group_layout_id: BindGroupLayoutId,
        descriptor: Option<BindGroupLayoutDescriptor<'static>>,
    },
    CreateBuffer {
        device_id: DeviceId,
        buffer_id: BufferId,
        descriptor: BufferDescriptor<'static>,
    },
    CreateCommandEncoder {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        desc: CommandEncoderDescriptor<Label<'static>>,
    },
    CreateComputePipeline {
        device_id: DeviceId,
        compute_pipeline_id: ComputePipelineId,
        descriptor: ComputePipelineDescriptor<'static>,
        async_sender: Option<GenericCallback<WebGPUComputePipelineResponse>>,
    },
    CreatePipelineLayout {
        device_id: DeviceId,
        pipeline_layout_id: PipelineLayoutId,
        descriptor: PipelineLayoutDescriptor<'static>,
    },
    CreateRenderPipeline {
        device_id: DeviceId,
        render_pipeline_id: RenderPipelineId,
        descriptor: RenderPipelineDescriptor<'static>,
        async_sender: Option<GenericCallback<WebGPURenderPipelineResponse>>,
    },
    CreateSampler {
        device_id: DeviceId,
        sampler_id: SamplerId,
        descriptor: SamplerDescriptor<'static>,
    },
    CreateShaderModule {
        device_id: DeviceId,
        program_id: ShaderModuleId,
        program: String,
        label: Option<String>,
        callback: GenericCallback<Option<ShaderCompilationInfo>>,
    },
    CreateContext {
        buffer_ids: ArrayVec<BufferId, PRESENTATION_BUFFER_COUNT>,
        size: DeviceIntSize,
        sender: GenericSender<WebGPUContextId>,
    },
    Present {
        context_id: WebGPUContextId,
        pending_texture: Option<PendingTexture>,
        size: Size2D<u32>,
        canvas_epoch: Epoch,
    },
    GetImage {
        context_id: WebGPUContextId,
        pending_texture: Option<PendingTexture>,
        sender: GenericSender<SharedSnapshot>,
    },
    ValidateTextureDescriptor {
        device_id: DeviceId,
        texture_id: TextureId,
        descriptor: TextureDescriptor<'static>,
    },
    DestroyContext {
        context_id: WebGPUContextId,
    },
    CreateTexture {
        device_id: DeviceId,
        texture_id: TextureId,
        descriptor: TextureDescriptor<'static>,
    },
    CreateTextureView {
        texture_id: TextureId,
        texture_view_id: TextureViewId,
        device_id: DeviceId,
        descriptor: Option<TextureViewDescriptor<'static>>,
    },
    DestroyBuffer(BufferId),
    DestroyDevice(DeviceId),
    DestroyTexture(TextureId),
    DropTexture(TextureId),
    DropAdapter(AdapterId),
    DropDevice(DeviceId),
    DropBuffer(BufferId),
    DropPipelineLayout(PipelineLayoutId),
    DropComputePipeline(ComputePipelineId),
    DropRenderPipeline(RenderPipelineId),
    DropBindGroup(BindGroupId),
    DropBindGroupLayout(BindGroupLayoutId),
    DropCommandEncoder(CommandEncoderId),
    DropCommandBuffer(CommandBufferId),
    DropTextureView(TextureViewId),
    DropSampler(SamplerId),
    DropShaderModule(ShaderModuleId),
    DropRenderBundle(RenderBundleId),
    DropQuerySet(QuerySetId),
    DropComputePass(ComputePassEncoderId),
    DropRenderPass(RenderPassEncoderId),
    Exit(GenericOneshotSender<()>),
    RenderBundleEncoderFinish {
        render_bundle_encoder_id: RenderBundleEncoderId,
        descriptor: RenderBundleDescriptor<'static>,
        render_bundle_id: RenderBundleId,
        device_id: DeviceId,
    },
    RequestAdapter {
        sender: GenericCallback<WebGPUAdapterResponse>,
        options: RequestAdapterOptions,
        adapter_id: AdapterId,
    },
    RequestDevice {
        sender: GenericCallback<WebGPUDeviceResponse>,
        adapter_id: WebGPUAdapter,
        descriptor: DeviceDescriptor<Option<String>>,
        device_id: DeviceId,
        queue_id: QueueId,
        pipeline_id: PipelineId,
    },
    BeginComputePass {
        command_encoder_id: CommandEncoderId,
        compute_pass_id: ComputePassId,
        label: Label<'static>,
        timestamp_writes: Option<PassTimestampWrites>,
        device_id: DeviceId,
    },
    ComputePassSetPipeline {
        compute_pass_id: ComputePassId,
        pipeline_id: ComputePipelineId,
        device_id: DeviceId,
    },
    ComputePassSetBindGroup {
        compute_pass_id: ComputePassId,
        index: u32,
        bind_group_id: BindGroupId,
        offsets: Vec<u32>,
        device_id: DeviceId,
    },
    ComputePassDispatchWorkgroups {
        compute_pass_id: ComputePassId,
        x: u32,
        y: u32,
        z: u32,
        device_id: DeviceId,
    },
    ComputePassDispatchWorkgroupsIndirect {
        compute_pass_id: ComputePassId,
        buffer_id: BufferId,
        offset: u64,
        device_id: DeviceId,
    },
    ComputePassPushDebugGroup {
        compute_pass_id: ComputePassId,
        label: String,
        device_id: DeviceId,
    },
    ComputePassPopDebugGroup {
        compute_pass_id: ComputePassId,
        device_id: DeviceId,
    },
    ComputePassInsertDebugMarker {
        compute_pass_id: ComputePassId,
        label: String,
        device_id: DeviceId,
    },
    EndComputePass {
        compute_pass_id: ComputePassId,
        device_id: DeviceId,
    },
    BeginRenderPass {
        command_encoder_id: CommandEncoderId,
        render_pass_id: RenderPassId,
        label: Label<'static>,
        color_attachments: Vec<Option<RenderPassColorAttachment>>,
        depth_stencil_attachment: Option<RenderPassDepthStencilAttachment<TextureViewId>>,
        timestamp_writes: Option<PassTimestampWrites>,
        device_id: DeviceId,
    },
    RenderPassCommand {
        render_pass_id: RenderPassId,
        render_command: RenderCommand,
        device_id: DeviceId,
    },
    EndRenderPass {
        render_pass_id: RenderPassId,
        device_id: DeviceId,
    },
    Submit {
        device_id: DeviceId,
        queue_id: QueueId,
        command_buffers: Vec<CommandBufferId>,
    },
    UnmapBuffer {
        buffer_id: BufferId,
        mapping: Option<Mapping>,
    },
    WriteBuffer {
        device_id: DeviceId,
        queue_id: QueueId,
        buffer_id: BufferId,
        buffer_offset: u64,
        data: GenericSharedMemory,
    },
    WriteTexture {
        device_id: DeviceId,
        queue_id: QueueId,
        texture_cv: TexelCopyTextureInfo,
        data_layout: TexelCopyBufferLayout,
        size: Extent3d,
        data: GenericSharedMemory,
    },
    QueueOnSubmittedWorkDone {
        sender: GenericCallback<()>,
        queue_id: QueueId,
    },
    PushErrorScope {
        device_id: DeviceId,
        filter: ErrorFilter,
    },
    DispatchError {
        device_id: DeviceId,
        error: Error,
    },
    PopErrorScope {
        device_id: DeviceId,
        callback: GenericCallback<WebGPUPoppedErrorScopeResponse>,
    },
    ComputeGetBindGroupLayout {
        device_id: DeviceId,
        pipeline_id: ComputePipelineId,
        index: u32,
        id: BindGroupLayoutId,
    },
    RenderGetBindGroupLayout {
        device_id: DeviceId,
        pipeline_id: RenderPipelineId,
        index: u32,
        id: BindGroupLayoutId,
    },
    CreateQuerySet {
        device_id: DeviceId,
        query_set_id: QuerySetId,
        descriptor: QuerySetDescriptor<'static>,
    },
    ResolveQuerySet {
        device_id: DeviceId,
        command_encoder_id: CommandEncoderId,
        query_set_id: QuerySetId,
        start_query: u32,
        query_count: u32,
        destination: BufferId,
        destination_offset: u64,
    },
    CreatePlanarTexture {
        device_id: DeviceId,
        size: Size2D<u32>,
        format: SnapshotPixelFormat,
        texture_id: TextureId,
        texture_view_id: TextureViewId,
    },
    UpdatePlanarTexture {
        device_id: DeviceId,
        queue_id: QueueId,
        texture_id: TextureId,
        snapshot: SharedSnapshot,
    },
    DropPlanarTexture(TextureId, TextureViewId),
    ImportExternalTexture {
        device_id: DeviceId,
        external_texture_id: ExternalTextureId,
        label: String,
        size: Size2D<u32>,
        plane0: Option<TextureViewId>,
    },
    DestroyExternalTexture(ExternalTextureId),
    DropExternalTexture(ExternalTextureId),
    DestroyQuerySet(QuerySetId),
    CreateRenderBundleEncoder {
        device_id: DeviceId,
        render_bundle_encoder_id: RenderBundleEncoderId,
        desc: RenderBundleEncoderDescriptor<'static>,
    },
    RenderBundleEncoderCommand {
        render_bundle_encoder_id: RenderBundleEncoderId,
        render_command: RenderBundleCommand,
        device_id: DeviceId,
    },
    DropRenderBundleEncoder(RenderBundleEncoderId),

}

Variants

SetImageKey

Fields

context_id: WebGPUContextId

image_key: ImageKey

BufferMapAsync

Fields

callback: GenericCallback<Result<Mapping, BufferAccessError>>

buffer_id: BufferId

device_id: DeviceId

host_map: HostMap

offset: u64

size: Option<u64>

CommandEncoderFinish

Fields

command_encoder_id: CommandEncoderId

device_id: DeviceId

desc: CommandBufferDescriptor<Label<'static>>

command_buffer_id: CommandBufferId

CopyBufferToBuffer

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

source_id: BufferId

source_offset: BufferAddress

destination_id: BufferId

destination_offset: BufferAddress

size: BufferAddress

CopyBufferToTexture

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

source: TexelCopyBufferInfo

destination: TexelCopyTextureInfo

copy_size: Extent3d

CopyTextureToBuffer

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

source: TexelCopyTextureInfo

destination: TexelCopyBufferInfo

copy_size: Extent3d

CopyTextureToTexture

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

source: TexelCopyTextureInfo

destination: TexelCopyTextureInfo

copy_size: Extent3d

CopyExternalImageToTexture

Fields

device_id: DeviceId

queue_id: QueueId

usable_source: Option<SharedSnapshot>

destination: TexelCopyTextureInfo

dest_tex_descriptor: TextureDescriptor<'static>

copy_size: Extent3d

CommandEncoderPushDebugGroup

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

label: String

CommandEncoderPopDebugGroup

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

CommandEncoderInsertDebugMarker

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

label: String

CreateBindGroup

Fields

device_id: DeviceId

bind_group_id: BindGroupId

descriptor: BindGroupDescriptor<'static>

CreateBindGroupLayout

Fields

device_id: DeviceId

bind_group_layout_id: BindGroupLayoutId

descriptor: Option<BindGroupLayoutDescriptor<'static>>

CreateBuffer

Fields

device_id: DeviceId

buffer_id: BufferId

descriptor: BufferDescriptor<'static>

CreateCommandEncoder

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

desc: CommandEncoderDescriptor<Label<'static>>

CreateComputePipeline

Fields

device_id: DeviceId

compute_pipeline_id: ComputePipelineId

descriptor: ComputePipelineDescriptor<'static>

async_sender: Option<GenericCallback<WebGPUComputePipelineResponse>>

present only on ASYNC versions

CreatePipelineLayout

Fields

device_id: DeviceId

pipeline_layout_id: PipelineLayoutId

descriptor: PipelineLayoutDescriptor<'static>

CreateRenderPipeline

Fields

device_id: DeviceId

render_pipeline_id: RenderPipelineId

descriptor: RenderPipelineDescriptor<'static>

async_sender: Option<GenericCallback<WebGPURenderPipelineResponse>>

present only on ASYNC versions

CreateSampler

Fields

device_id: DeviceId

sampler_id: SamplerId

descriptor: SamplerDescriptor<'static>

CreateShaderModule

Fields

device_id: DeviceId

program_id: ShaderModuleId

program: String

label: Option<String>

callback: GenericCallback<Option<ShaderCompilationInfo>>

CreateContext

Creates context

Fields

buffer_ids: ArrayVec<BufferId, PRESENTATION_BUFFER_COUNT>

size: DeviceIntSize

sender: GenericSender<WebGPUContextId>

Present

Present texture to WebRender

Fields

context_id: WebGPUContextId

pending_texture: Option<PendingTexture>

size: Size2D<u32>

canvas_epoch: Epoch

GetImage

Create pixels::Snapshot with contents of the last present operation or provided pending texture and send it over provided [IpcSender].

Fields

context_id: WebGPUContextId

pending_texture: Option<PendingTexture>

sender: GenericSender<SharedSnapshot>

ValidateTextureDescriptor

Fields

device_id: DeviceId

texture_id: TextureId

descriptor: TextureDescriptor<'static>

DestroyContext

Fields

context_id: WebGPUContextId

CreateTexture

Fields

device_id: DeviceId

texture_id: TextureId

descriptor: TextureDescriptor<'static>

CreateTextureView

Fields

texture_id: TextureId

texture_view_id: TextureViewId

device_id: DeviceId

descriptor: Option<TextureViewDescriptor<'static>>

DestroyBuffer(BufferId)

DestroyDevice(DeviceId)

DestroyTexture(TextureId)

DropTexture(TextureId)

DropAdapter(AdapterId)

DropDevice(DeviceId)

DropBuffer(BufferId)

DropPipelineLayout(PipelineLayoutId)

DropComputePipeline(ComputePipelineId)

DropRenderPipeline(RenderPipelineId)

DropBindGroup(BindGroupId)

DropBindGroupLayout(BindGroupLayoutId)

DropCommandEncoder(CommandEncoderId)

DropCommandBuffer(CommandBufferId)

DropTextureView(TextureViewId)

DropSampler(SamplerId)

DropShaderModule(ShaderModuleId)

DropRenderBundle(RenderBundleId)

DropQuerySet(QuerySetId)

DropComputePass(ComputePassEncoderId)

DropRenderPass(RenderPassEncoderId)

Exit(GenericOneshotSender<()>)

RenderBundleEncoderFinish

Fields

render_bundle_encoder_id: RenderBundleEncoderId

descriptor: RenderBundleDescriptor<'static>

render_bundle_id: RenderBundleId

device_id: DeviceId

RequestAdapter

Fields

sender: GenericCallback<WebGPUAdapterResponse>

options: RequestAdapterOptions

adapter_id: AdapterId

RequestDevice

Fields

sender: GenericCallback<WebGPUDeviceResponse>

adapter_id: WebGPUAdapter

descriptor: DeviceDescriptor<Option<String>>

device_id: DeviceId

queue_id: QueueId

pipeline_id: PipelineId

BeginComputePass

Fields

command_encoder_id: CommandEncoderId

compute_pass_id: ComputePassId

label: Label<'static>

timestamp_writes: Option<PassTimestampWrites>

device_id: DeviceId

ComputePassSetPipeline

Fields

compute_pass_id: ComputePassId

pipeline_id: ComputePipelineId

device_id: DeviceId

ComputePassSetBindGroup

Fields

compute_pass_id: ComputePassId

index: u32

bind_group_id: BindGroupId

offsets: Vec<u32>

device_id: DeviceId

ComputePassDispatchWorkgroups

Fields

compute_pass_id: ComputePassId

x: u32

y: u32

z: u32

device_id: DeviceId

ComputePassDispatchWorkgroupsIndirect

Fields

compute_pass_id: ComputePassId

buffer_id: BufferId

offset: u64

device_id: DeviceId

ComputePassPushDebugGroup

Fields

compute_pass_id: ComputePassId

label: String

device_id: DeviceId

ComputePassPopDebugGroup

Fields

compute_pass_id: ComputePassId

device_id: DeviceId

ComputePassInsertDebugMarker

Fields

compute_pass_id: ComputePassId

label: String

device_id: DeviceId

EndComputePass

Fields

compute_pass_id: ComputePassId

device_id: DeviceId

BeginRenderPass

Fields

command_encoder_id: CommandEncoderId

render_pass_id: RenderPassId

label: Label<'static>

color_attachments: Vec<Option<RenderPassColorAttachment>>

depth_stencil_attachment: Option<RenderPassDepthStencilAttachment<TextureViewId>>

timestamp_writes: Option<PassTimestampWrites>

device_id: DeviceId

RenderPassCommand

Fields

render_pass_id: RenderPassId

render_command: RenderCommand

device_id: DeviceId

EndRenderPass

Fields

render_pass_id: RenderPassId

device_id: DeviceId

Submit

Fields

device_id: DeviceId

queue_id: QueueId

command_buffers: Vec<CommandBufferId>

UnmapBuffer

Fields

buffer_id: BufferId

mapping: Option<Mapping>

Return back mapping for writeback

WriteBuffer

Fields

device_id: DeviceId

queue_id: QueueId

buffer_id: BufferId

buffer_offset: u64

data: GenericSharedMemory

WriteTexture

Fields

device_id: DeviceId

queue_id: QueueId

texture_cv: TexelCopyTextureInfo

data_layout: TexelCopyBufferLayout

size: Extent3d

data: GenericSharedMemory

QueueOnSubmittedWorkDone

Fields

sender: GenericCallback<()>

queue_id: QueueId

PushErrorScope

Fields

device_id: DeviceId

filter: ErrorFilter

DispatchError

Fields

device_id: DeviceId

error: Error

PopErrorScope

Fields

device_id: DeviceId

callback: GenericCallback<WebGPUPoppedErrorScopeResponse>

ComputeGetBindGroupLayout

Fields

device_id: DeviceId

pipeline_id: ComputePipelineId

index: u32

id: BindGroupLayoutId

RenderGetBindGroupLayout

Fields

device_id: DeviceId

pipeline_id: RenderPipelineId

index: u32

id: BindGroupLayoutId

CreateQuerySet

Fields

device_id: DeviceId

query_set_id: QuerySetId

descriptor: QuerySetDescriptor<'static>

ResolveQuerySet

Fields

device_id: DeviceId

command_encoder_id: CommandEncoderId

query_set_id: QuerySetId

start_query: u32

query_count: u32

destination: BufferId

destination_offset: u64

CreatePlanarTexture

Create planar texture and view to be imported as external texture

Fields

device_id: DeviceId

size: Size2D<u32>

format: SnapshotPixelFormat

texture_id: TextureId

texture_view_id: TextureViewId

aka plane

UpdatePlanarTexture

Fields

device_id: DeviceId

queue_id: QueueId

texture_id: TextureId

snapshot: SharedSnapshot

DropPlanarTexture(TextureId, TextureViewId)

ImportExternalTexture

Import plane as external texture, if plane not provided it creates invalid external texture

Fields

device_id: DeviceId

external_texture_id: ExternalTextureId

label: String

size: Size2D<u32>

plane0: Option<TextureViewId>

DestroyExternalTexture(ExternalTextureId)

DropExternalTexture(ExternalTextureId)

DestroyQuerySet(QuerySetId)

CreateRenderBundleEncoder

Fields

device_id: DeviceId

render_bundle_encoder_id: RenderBundleEncoderId

desc: RenderBundleEncoderDescriptor<'static>

RenderBundleEncoderCommand

Fields

render_bundle_encoder_id: RenderBundleEncoderId

render_command: RenderBundleCommand

device_id: DeviceId

DropRenderBundleEncoder(RenderBundleEncoderId)